+++ lib/services/walletconnect_modal/walletconnect_modal_service.dart
@@ -70,6 +70,9 @@ class WalletConnectModalService extends ChangeNotifier
   @override
   Map<String, RequiredNamespace> get optionalNamespaces => _optionalNamespaces;
 
+  @override
+  final Event<EventArgs> onPairingExpireEvent = Event();
+
   ConnectResponse? connectResponse;
   Future<SessionData>? get sessionFuture => connectResponse?.session.future;
   BuildContext? context;
@@ -136,10 +139,8 @@ class WalletConnectModalService extends ChangeNotifier
     _initError = null;
     try {
       await _web3App!.init();
-    } catch (_) {}
-
-    try {
       await WalletConnectModalServices.init();
+      await clearPreviousInactivePairings();
     } catch (_) {}
 
     if (_web3App!.sessions.getAll().isNotEmpty) {
@@ -153,6 +154,15 @@ class WalletConnectModalService extends ChangeNotifier
     _isInitialized = true;
 
     notifyListeners();
+  }
+
+  @override
+  Future<void> clearPreviousInactivePairings() async {
+    for (var pairing in _web3App!.pairings.getAll()) {
+      if (!pairing.active) {
+        await _web3App!.core.expirer.expire(pairing.topic);
+      }
+    }
   }
 
   @override
@@ -176,12 +186,8 @@ class WalletConnectModalService extends ChangeNotifier
 
     _isOpen = true;
 
-    rebuildConnectionUri();
-
     // Reset the explorer
-    explorerService.instance!.filterList(
-      query: '',
-    );
+    explorerService.instance!.filterList(query: '');
     widgetStack.instance.clear();
 
     this.context = context;
@@ -341,15 +347,20 @@ class WalletConnectModalService extends ChangeNotifier
         }
       }
 
+      _web3App!.core.pairing.onPairingExpire.subscribe(_onPairingExpire);
       connectResponse = await web3App!.connect(
         requiredNamespaces: requiredNamespaces,
         optionalNamespaces: optionalNamespaces,
       );
 
       notifyListeners();
-
       awaitConnectResponse();
     }
+  }
+
+  void _onPairingExpire(PairingEvent? args) async {
+    _web3App!.core.pairing.onPairingExpire.unsubscribe(_onPairingExpire);
+    onPairingExpireEvent.broadcast();
   }
 
   bool _connectingWallet = false;