+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.12
+
+- Update `walletconnect_flutter_v2` to use latest 2.1.5 version
+- Rebuild connection URI after pairing expiration.
+  
 ## 2.1.11
 
 - `WalletConnectModalService.disconnect` now disconnects all sessions, unless you tell it to only disconnect the current one